Set up documentation site
The prompt products data documentation site is based on the same
framework as dp1.lsst.io. Is it also configured to not immediately
upload. By changing the .github/workflows/ci.yaml workflow the uploads
to prompt-products.lsst.io can be toggled on. In the meantime, the built
HTML is available as a downloadable artifact from the Actions tab of the
repository on GitHub.